Create predict_proba for ensemble in predict
